move lib js files to lib directory (to be cleaner)
