+++ lib/cleanup.js
