+++ lib/imgManip.js
