+++ lib/reset.js
