+++ lib/setHeaders.js
