+++ lib/spriteMaker.js
