+++ test/post/index.js
@@ -11,4 +11,6 @@ describe('post', function(){
 
     })
 
+  })
+
 })
